+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.7.0] - 2024-04-15
+
+### Added
+
+- Added optional serialization methods `to_json`, `to_speedy` and `to_bincode` to `Walk`. The corresponding features `json`, `speedy` and `bincode` need to be enabled.
+- Add `statistics` getter to `Walk`.
+
+### Improved
+
+- Optimized code.
+- Update benchmarks.
+
 ## [2.6.0] - 2024-04-10
 
 ### Added